requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
alembic==1.14.0
    # via -r requirements.in
annotated-types==0.7.0
    # via
    #   -r requirements.in
    #   pydantic
anyio==4.7.0
    # via
    #   -r requirements.in
    #   starlette
bcrypt==4.2.1
    # via -r requirements.in
build==1.2.2.post1
    # via
    #   -r requirements.in
    #   pip-tools
cffi==1.17.1
    # via
    #   -r requirements.in
    #   cryptography
click==8.1.8
    # via
    #   -r requirements.in
    #   pip-tools
    #   uvicorn
cryptography==44.0.0
    # via
    #   -r requirements.in
    #   jwt
dnspython==2.7.0
    # via
    #   -r requirements.in
    #   email-validator
ecdsa==0.19.0
    # via
    #   -r requirements.in
    #   python-jose
email-validator==2.2.0
    # via -r requirements.in
exceptiongroup==1.2.2
    # via
    #   -r requirements.in
    #   anyio
fastapi==0.115.6
    # via -r requirements.in
gevent==24.11.1
    # via
    #   -r requirements.in
    #   websocket
greenlet==3.1.1
    # via
    #   -r requirements.in
    #   gevent
    #   sqlalchemy
    #   websocket
h11==0.14.0
    # via
    #   -r requirements.in
    #   uvicorn
idna==3.10
    # via
    #   -r requirements.in
    #   anyio
    #   email-validator
jose==1.0.0
    # via -r requirements.in
jwt==1.3.1
    # via -r requirements.in
mako==1.3.8
    # via
    #   -r requirements.in
    #   alembic
markupsafe==3.0.2
    # via
    #   -r requirements.in
    #   mako
packaging==24.2
    # via
    #   -r requirements.in
    #   build
passlib==1.7.4
    # via -r requirements.in
pip-tools==7.4.1
    # via -r requirements.in
psycopg2==2.9.10
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   -r requirements.in
    #   python-jose
    #   rsa
pycparser==2.22
    # via
    #   -r requirements.in
    #   cffi
pydantic==2.10.4
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.27.2
    # via
    #   -r requirements.in
    #   pydantic
pyproject-hooks==1.2.0
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
python-dotenv==1.0.1
    # via -r requirements.in
python-jose==3.3.0
    # via -r requirements.in
python-multipart==0.0.20
    # via -r requirements.in
rsa==4.9
    # via
    #   -r requirements.in
    #   python-jose
six==1.17.0
    # via
    #   -r requirements.in
    #   ecdsa
sniffio==1.3.1
    # via
    #   -r requirements.in
    #   anyio
sqlalchemy==2.0.36
    # via
    #   -r requirements.in
    #   alembic
starlette==0.41.3
    # via
    #   -r requirements.in
    #   fastapi
tomli==2.2.1
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   uvicorn
uvicorn==0.34.0
    # via -r requirements.in
websocket==0.2.1
    # via -r requirements.in
wheel==0.45.1
    # via pip-tools
zope-event==5.0
    # via
    #   -r requirements.in
    #   gevent
zope-interface==7.2
    # via
    #   -r requirements.in
    #   gevent

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
